Added skip_node_ops flag to app/domain/user models.
Add skip_node_ops to District model

Added 'fix-unresponsive-apps' option to oo-admin-repair script

+++ controller/app/models/pending_ops/destroy_gear_op.rb
@@ -3,9 +3,12 @@ class DestroyGearOp < PendingAppOp
   field :gear_id, type: String
   field :group_instance_id, type: String
 
-  def execute()
-    gear = get_gear()
-    result_io = gear.destroy_gear(true)
+  def execute(skip_node_ops=false)
+    result_io = ResultIO.new
+    unless skip_node_ops
+      gear = get_gear()
+      result_io = gear.destroy_gear(true)
+    end
     result_io
   end
